[front] Move editForm components
