+++ next_app/src/features/room_video/components/admin/UpdateContentForm.tsx
@@ -3,14 +3,14 @@ import { useEffect, useState } from 'react'
 import { useRouter } from 'next/navigation'
 import axios from 'axios'
 
-import css from 'components/editForm/EditContentForm.module.scss'
-import { useEditDescription } from 'components/editForm/EditDescription'
-import { useEditImage } from 'components/editForm/EditImage'
-import { useEditTags } from 'components/editForm/EditTags'
-import { useEditTitle } from 'components/editForm/EditTitle'
-import { useEditVideo } from 'components/editForm/EditVideo'
-import { useProgressBar } from 'components/editForm/ProgressBar'
-import ContentDeleteButton from 'components/editForm/ContentDeleteButton'
+import css from 'features/rooms/components/editForm/EditContentForm.module.scss'
+import { useEditDescription } from 'features/rooms/components/editForm/EditDescription'
+import { useEditImage } from 'features/rooms/components/editForm/EditImage'
+import { useEditTags } from 'features/rooms/components/editForm/EditTags'
+import { useEditTitle } from 'features/rooms/components/editForm/EditTitle'
+import { useEditVideo } from 'features/rooms/components/editForm/EditVideo'
+import { useProgressBar } from 'features/rooms/components/editForm/ProgressBar'
+import ContentDeleteButton from 'features/rooms/components/editForm/ContentDeleteButton'
 import PopupWindowMessage from 'components/PopupWindowMessage'
 import SimpleButton from 'components/SimpleButton'
 import { TagData } from 'features/tags/tags'
